Initial changes to test the controller.proto seam
o Important Note: I sadly added a dependency from aerofs.gui on aerofs.spsv.
The LocalSPServiceReactorCaller is in the com.aerofs.sp.server.sp package
as it depends on SPService and other SP classes. The TestControllerSetup class
is in the com.aerofs.controller package as it instantiates a controller.
Unfortunately TestControllerSetup has a dependency on the
LocalSPServiceReactorCaller, and that's where the module dependency comes in.
Any suggestions? I'm not thrilled about it.

o one test successfully verifies that ExNotFound is thrown
when an invitation code does not exist

o added test configuration for TestSPACL and TestControllerSetup (running from
approot)

Change-Id: I0536b908f2bc0aa170fd1d5fd23e6aa1c24f9c93