Add S3 libs/classes to build
Modified the build script to copy the s3 libs to the linux dist dir and
add classpath references for them. Moved the libs from
java/aerofs.s3/lib to as-is/s3/lib for consistency. Also added some
targets to the ant build.xml file to build the client jar and run
ProGuard to obfuscate it, but these aren't used anywhere yet. It seems
like it would be good to minimize the different places things are built
but I don't have a strong opinion.

Change-Id: Ibe07f814a4ee86dc410c8417b7fb0feae1937750