Finsh client-verkehr side of mutal auth using CRLs.
More specifically...
- Update certify device so that the serial and expire timestamp are updated in
  the SP database when a cert is generated.
- Add new revoke device/user certificate SP calls (and unit tests).
- Hook up a verkehr command channel that instructs verkehr to pull the
  new updated CRL (and create some associated verkehr example code).
- Enforce auth using CRLs on the verkehr side.

Need to release a new verkehr, sp, client, and update the db all at the same
time when releasing this commit (db change is just adding a new table).

Client to client auth using CRLs is not finished yet. Need to finish up
the verkehr client notifications on the SP side and also need to update
the client code (deferred to a later date...).

Change-Id: I0cac30737b03be462e2e3d67e75f53fd4919e53e