Reexecute aerofs-cli on Linux, even if not in PATH
After updates, we're very good at killing aerofs-cli, but not so great
at starting it again.  In particular, if a user uses the .tgz package to
install aerofs, then aerofs-cli is not necessarily on the user's PATH,
so when updater.sh executes aerofs-cli, it winds up launching nothing.

This patch makes the Linux packages include the launcher script, and
makes the updater script run the commands from .aerofs-bin.

Change-Id: I071a827f73a9989cd33a98a9e6e1c1afd4a442d5