Mobile client support
The plan is for the desktop clients to support an RPC interface for
mobile clients to access that allows the latter to list and download
files. Added a mobile service on its own port that supports this
interface. Right now it's not authenticated at all, but I'm going to
change it to use TLS.

Change-Id: If264b72f5d1684af237c3ab39a9fc0e78b148671