Canonicalize root anchor path when read from cfg db
On OSX, all file notifications are delivered with canonical paths.
However, Path has to use abstract pathnames to determine if one Path is
a child of another path, because there's no guarantee that the two paths
exist on the filesystem.

Thus, we need to track all paths that may receive OS watch notifications
as canonical paths.  However, to reduce user confusion, we wish to keep
the user-provided path in the config DB, so we simply canonicalize the
path when the field is read from the DB.

Change-Id: Ia1aaed191e9d46303bae1c69b55bbf6732e8bfd0
Addresses: aerofs-859