Remove an overzealous assertion.
This assertion can be triggered when a file gets changed after the
daemon has read the whole file but is still uploading the last chunk.
The remote peer receives all correct data, and if we don't crash the
daemon, then we'll pick up the change from the notifier later and handle
it correctly.  This codepath is probably triggerable once in a blue moon
if you have two peers syncing while you populate your root anchor with
rsync or something.

Change-Id: I15d37076529417f476ce96485340c46e08fd12b4
Addresses: aerofs-1156