aerofsd: a JNI-based launcher for the AeroFS daemon
The primary benefit of this work is that users can now set policy on
aerofsd specifically, rather than on all Java applications as a whole.

liblauncher required significant refactoring to work correctly with
UTF16 paths on Windows.  Even so, Java's System.load() doesn't seem to
handle such filenames correctly, so there's no guarantee this will work
when there are non-ASCII characters in the path to the approot.

Additionally, we ensure that aerofsd always changes its pwd to be the
approot before launching to guarantee that it will find its native
libraries.

Change-Id: I34fc420ae8673e4944baae28956785a0e7b52ffc