GetHash -> ComputeHashCall
When we encounter an aliased file with content present on both the
remote and local peer, we compare ContentHash to see if the files are in
fact identical.  However, we request (and compute!) this hash in the
middle of receiving the file contents, which get queued until we exhaust
available heap space.

To avoid this outcome, we change the protocol from having a "GetHash"
call which gets the hash content to having an advisory "ComputeHash"
call.  Then, if we need hashes to compare file contents, we simply abort
the transfer, send the advisory "compute the hash for this version"
message to the peer, and then rerequest the download.

(This patchset only fixes the OOM; stopping excess file download will
happen in a separate patchset)

Change-Id: I3a37a2eb2ae96c9e31b0ded8423985f493b46b1c