Find JRE path even if JDK 'java' executable is on $PATH
Oracle does an annoying thing and distributes the JDK package with a
'java' executable in bin/java and another in jre/bin/java.  We
previously assumed that it was the JRE's executable that was on the
user's PATH, but some users (and developers :P) have the JDK installed
and wind up using that java binary.

This patchset allows aerofsd to find its required native libraries in
such a situation.

Change-Id: Ieb379734081b535c3ee8263689cb0698c57d445e