aerofsd: use the default codepage on Windows, not UTF8.
While my dive into the OpenJDK source led me to believe that all
optionStrings were supposed to be encoded as UTF8, the JVM fails to find
the classes in the aerofs jar specified on the classpath if you are on
codepage 1252 and have a username that uses a high-bit Latin-1
character.

I'd rather be making the JVM interpret that path as a UTF-8 string, but
a day knee-deep in JVM code has left me hopeless, so I'm giving up and
just submitting this.

Change-Id: Ibe6c78f28a3eb2614a9ce1d338e1b721200ceb1f
Fixes: support-156
Fixes: support-157