Fix possible deadlock in Smack initialization
I ran into a deadlock when using a local ejabberd xmpp server for
testing: there's a race condition between two threads that both run
static class initializers, the thread making the xmpp connection and the
thread adding the stores and thus trying to join mucs. This should fix
the issue.

Change-Id: Ie65b8fd98d338f66599aba164f50d7788f34afd0