SyncStatusConnection: unresponsive server => DISCONNECTED
The syncstat server will accept incoming connections even though it is overloaded and
cannot service it. Which result in the status switching to DISCONNECTED on every RPC
timeout and immediately back to CONNECTED at the next attempt, falsely leading the
user to believe that the sync status information is accurate.

Change-Id: Iab54386014a5b9b7ccf718bf997d9ed7e636780c