Dump heap to rtroot on OOM and send hprof to SV
This patchset adds logic to aerofsd to set the HeapDumpPath to the
rtroot (which we know to be writeable) and to have SVClient send the
hprof files along with the logfiles.

This allows us to automatically collect heap dumps when users hit OOM
errors, which should help us diagnose the issues.

Change-Id: I178c23e5fa30af920656fcb37a1aed88027a3d7a