Update to OpenSSL 1.0.1c.
There was a security vulnerability in OpenSSL's DTLS implementation that
could lead to plaintext recovery:
<http://www.isg.rhul.ac.uk/~kp/dtls.pdf>

Since libaerofsd links to OpenSSL, we must also rebuild libaerofsd for
all platforms.

Rather than continue to ship a copy of libcrypto.so.1.0.0 for the python
libraries, we shall instead make aerofs server debian packages depend on
libssl1.0.0 and let the OS provide it for us (we're using the stock one
anyway).  Interestingly, the library in the admin panel appears unused
in prod anyway.

This changeset also gets rid of all duplicate copies of libcrypto and
libssl.  Everything is now a symbolic link to the official copy in
resource/client.

Change-Id: I63ec73031c73b9b439d6f7a85467bd41813316d8