Bump pagerduty url timeout to 3 seconds.
After doing some testing on the syncstat servlet while it was under
heavy load I observed that the average GET response time is about 1.3
seconds. This latency will cause a pagerduty false negative with the
existing 1 second timeout, so bump to 3 (to be safe). With the longer
timeout, the probe is still a good measure of health for all servers.

Change-Id: I19b7c9a0f7de816db54d5cd06afd62fd6c0c5ce8