Increase log verbosity in SyncStatusSynchronizer
core.syncstatus.should_report_migrated_files_synced had a weird failure (i.e. unreproducible)
It seems to be related to conversion of folder into store and based on a (long, way too damn
long) analysis of db and daemon log of the failing actor it does not appear to be an aggregation
bug (which is fortunate as the sync status DPUT has already been pushed) but rather a subtly
wrong sync status.

Increase log verbosity to get a clearer picture in case we ever run into this failure mode
again.

Change-Id: I1d9fbbdf89da5f45e6d21c4cee5f04c391bc1793