OSUtilLinux: determine filesystem type with stat; support vzfs
The GNU coreutils are more likely to know the filesystem superblock type
than our manually-generated magic table.  So shell out to stat on Linux
to detect the filesystem backing a path, and move the native call from
AbstractOSUtilLinuxOSX to OSUtilOSX (which still uses a native call).

Tested on Linux and OSX, the two affected platforms.

This commit *actually* adds support for VZFS.

Change-Id: I7d3f8c3be21f53dbc2995aac51d2797d87902bbd