SPDatabase: always give the ID present in the DB
For some cases where we looked up emails and organization IDs in the
database, we'd return a User or Organization with the string the user
passed us, rather than the one that the database contains.  As a result,
there were some issues relating to case insensitivity when an uppercased
query for a email in the DB as lowercase returned a user in uppercase.

Case-insensitivity is a pain.

Change-Id: Ic87a829c4a84f735cfb8c10cbd7557f4f2ac1698