Temporarily enable strict consistency checking in AggregateSyncStatus
A couple users are hitting assertion errors, unfortuntely it happens long after
the initial inconsistency appeared in the database which makes it extremely hard
to debug, especially given the absence of logging in many cases.

This consistency checking essentially negates the benefit of maintaining an
aggregate sync status and should therefore be disabled as soon as the root
cause(s) are found and fixed.

Change-Id: I0dc3e5bb7463138f594a89f7d33ee4dff6a4d519