Fix sign-in handler
Refactor sign-in handling to use exception instead of scheduling
an event for later processing (needed to safely reset pull epoch).

Reset pull epoch on every sign-in to avoid accidentally ignoring
incoming notifications in case of server-side data loss.

Start a new scan in case of push epoch rollback when no scan is
ongoing to make sure the vh lost by the server are swiftly re-sent.

Change-Id: Ia0d0c18d559d000297b668e8243409cb7bd917f8