Move Core and S3 DB schemas where they belong
Nothing outside the Core needs to know what these schema look like, in particular
it is seriously bad to have the controller explicitly setup the schemas. Instead
the daemon should automatically setup the schemas when the database file is missing.

In a next installment we should try to use injection to make the actual DB setup
logic cleaner.

Change-Id: I4122fedf9d06f532008b2bc259bd220fb9dd16f5