Rename IStoreDeletionListener to IStoreDeletionOperator and other related renames:
o StoreDeletionListenerNotifiers -> StoreDeletionOperators
o notifyListeners_ -> runAll_

Change-Id: Ief99dd6d691bc58c6af5c3ba07c5a77b556f7820
Reason: not all mediator/dispatcher patterns are necessarily the listener pattern. Listener patterns are by definition for event notifications and the publishing party should be fairly independent from the subscription party: The lack of any subsciber should not affect the publisher's consistency or correctness at all. In this case however, execution of the subscribers is critical to maintain correctness of the publisher. Therefore, to me, this situation doesn't fit the listener pattern.