Changes to sss icon overlay. Make their meaning more intuitive.
Let
    U = num users sharing a file or folder
    D = num devices (belonging to you) that share a file or folder

And similarly:
    u = num users in sync with a given file or folder
    d = num devices (belonging to you) in sync with a given file or
        folder

Then the definition of sync status is as follows:

For a file:
    Green:  d = D && u = U
    Yellow: 0 < d < D || 0 < u < U
    Red:    d = 0 && u = 0

For a directory:
    Green:  d = D && u = U
    Yellow: 0 < d < D || 0 < u < U
    Red:    d = 0 && u = 0

This definition is preferred because, intuitively, the icon overlay for
a directory can be thought of as a function written as:

    dir_icon = worst_colour(file[] child_files, dir[] child_dirs)

So we get cases that look like this:

    Red
        Green
        Yellow
        Yellow
        Red

    Yellow
        Green
        Yellow
        Green
        Green

Where previously we had the unintuitive:

    Red
        Green
        Yellow
        Green
        Green

Why would this be red!?

Further details will be provided in the blog... I'm not sure what the
intended behaviour was before. Since it wasn't clearly defined, the
desired behaviour probably bounced around a bunch...

IMO the syncstatus design doc as it is written already describes the
above "new behaviour" so I didn't change it.

N.B. this does not fix the case where we have:

Red
    Yellow (U1:in_sync  U2:out_sync)
    Yellow (U1:out_sync U2:in_sync)

Will address this in a different commit.

Change-Id: Id44d0b44ca349f0769c2246b87d0e9622dd78433