Wrap SWT's Program.launch; use xdg-open on Linux
xdg-open is the freedesktop.org-approved method for launching handler
programs on Linux.  It's installed by default on pretty much every
Linux system that has a GUI, and fixes the issues we have when KDE users
can't open files or URLs when SWT's Program can't load libgvfs because
it's not installed.

Tested by testing each of the GUIUtil.launch() invocations on a system
running GNOME and a system running KDE without GNOME installed.

Change-Id: Ib75adf021b42a51da84f76e7594fcb247b8bb407