Weaken overzealous store deletion assert
We had a slightly over-zealous assertion that would crash the daemon
when a store was deleted in the same millisecond as an object under the
store, which was reproduced by a syncdet testcase.

This weakens the assertion to verify that *if* a store is deleted, then
any objects modified under that store have had their Version set to
zero.

Change-Id: Ie8f3bb7415d6c9b8a6270c42a765ce939b5fe0bb