Change the way we add users to an organization.
Previously we would just move them to our organization via the
invite_user sp call. Now we invite them via an email with a link that
they can click on if they choose to accept the invitation. The user is
only moved when they click the link.

Updated the admin panel "manage users" page as well, add module for
acceptance of organization invitation.

Also renamed OrgID => OrganizationID for consistency across the
codebase.

Change-Id: Iab0d2899f95da15a4d2fd2de996a76c6194295c4