Changing the aerofs.aux location to always be in the same directory as the root anchor.
Previously, we would have 2 different strategies: we would either put it
under RT root if RT root was on the same mount point as Root Anchor, or
we would put it at the root directory of Root Anchor.

The latter case turned to often be an issue, since the root directory
is often non-writable.

Change-Id: I1a321877f9fc90a9a6216540a9bb735008ae9faa