Client side of cert-based sign on.
Turns out changes to vhost/proxy.conf were no good! Need them in
vhost_autogen.conf. Made this change as well. As per comment in the code
the dependency of AbstractServletClientHandler on Cfg is not pretty but
IMO it's not worth the engineering effort to make it nice at this time.

This works nicely, tested with staging. Die staging, die.

Change-Id: I0c9c4951f7b056a4dc4570368b92f0b2e778d243