puppet: pagerduty probe for my.aerofs.com
Since this is now a production service that users require to be up to
access one of the core features of aerofs (joining stores), it would be
nice if we knew if it went down.

It's probably a good idea for us to eventually make this more
intelligent and actually sign in to verify that that works, but knowing
that the web server exists would be a good start.

Change-Id: I05a8daa2308072ac02e72a62d73486f09081cea9