Add autoexport to BlockStorage.
Autoexport is intended as a best-effort means to create a replica of the
current state of each store, but makes no strong guarantees.  It simply
passes writes to BlockStorage through to a folder on the filesystem.  As
a hint to the user that the export folder is a one-way datadump, files
are marked readonly.

UI for setting the config option and migration of the export folder will
come in future commits, but this demonstrates the feature.

Change-Id: I72509655e04515892608cab916b7643df89ed70e