Improve overall web ui workflow and page layout:
Use "span10" instead of "span12" for the whole page. The navigation bar always shows on the left with "span2", leaving the dashboard a "span8" to fill the rest of the page.

I added a style called "page blocks" for all dashboard pages. Page blocks are used to decorate "rows" or used standalone to make proper spacing between dashboard function groups.

To write one-page-block dashboards with a single column:

<h2>Headline</h2>
<p>Text</p>

To write multi-page-block dashboards with a single column:

<div class="page_block">
    <h2>Headline</h2>
    <p>Text</p>
</div>

To write multi-page-block dashboards with multiple columns:

<div class="row page_block">
    <div class="spanN">
        <h2>Headline</h2>
        <p>Text</p>
    </div>
    <div class="spanN">...</div>
</div>
<div class="row page_block">
    <div class="spanN">
        <h2>Headline</h2>
        <p>Text</p>
    </div>
    <div class="spanN">...</div>
</div>

Change-Id: I6d023c46d9af3e48f808d1c39f56c794ccb43aed