Propagate permanent errors along dependency chain
Collector makes a distinction between transient and permanent errors to avoid
repeatedly querying a peer about a component. However if the permanent error
is encountered along a dependency chain (CONTENT dep on META for instance) the
error was not correctly propagated. The component was therefore repeatedly
requested and the bloom filters were not cleared leading to significant bandwidth
usage, especially for users with a large number of ghost KMLs.

NB: I had to setup the test devices to reproduce this issue by surgically altering
the db. I tried and tried and tried, and ultimately failed to elicit this case
in a 2-device syncdet environment (extended w/ device reinstall mid-case) but it
appears that either more than 2 peers are required or the condition might have
arised from a bug in an old version.

Change-Id: I6a8a0c4f7f84af0eddeab37559dc4f32754b5dbf