Refuse to git-review if you haven't run unit tests
The hook just looks to see if the report file is generated; it's
possible that the file is stale or that you failed tests.  But at least
this should help remind people that they really ought to run tests
before getting things reviewed/merged.

You can always skip the checks with "git review -n master".

Change-Id: I30055ed63d8146ca1d16a0b1eb8c034796f12102