Make OWNER check consistent between deleteACL and leaveSharedFolder/ignoreInvitation
The new invariant is that there should always be at least one OWNER among the
members of the shared folder, regardless of their pending state. The idea is
to make sure their is always one user that can regain control of the share
and assume admin functions.

Also add unit tests for ignoreSharedFolderInvitation

Change-Id: If8c5c05e4258de32cc2b88edba079dce33c1e796