relocation: update autoexport folder config value
Since the default behavior is to set the autoexport dir to be the root
anchor, when we relocate the data at the root anchor, we also wind up
relocating the data in the autoexport folder.  Since the user probably
wanted to move both in such a scenario, we should update the config key.

It might make more sense (once we expose the autoexport folder as a
separate concept) if we check that the autoexport folder is *under* the
root anchor, and then update the path relative to the root anchor...but
that doesn't seem necessary at this point.

Change-Id: Iecf7b92ebd7a3669aa77e5cd66a1d88b3a29d92e