Demote anchors when migrating
When an anchor is migrated, the original anchor is deleted and a new one is
created. SharedFolderAutoLeaver would pick up the deletion and not the
creation, effectively turning any anchor migration into a deletion.

After talking to ww and comparing with Dropbox behavior it was decided to
change migration behavior instead of fixing SharedFolderautoLeaver:
migrating an anchor should convert it back to a regular folder

Change-Id: I0c205770844bf14c23b3b5a8beda910e1cc45346