SecUtil: Use an OpenSSL-backed PKCS10 implementation
(PKCS10 is the certificate signing request format standard)

sun.security.pkcs.PKCS10 is a JRE implementation detail, and is not
present in non-Sun JREs, nor Oracle Java 8.  We shouldn't depend on its
existence in the client, and since we already ship OpenSSL, we might as
well just use OpenSSL to provide a PKCS10 signer.

This commit does precisely that - it adds a module to aerofsd that
provides a PKCS10 implementation and in doing so breaks the daemon's
dependency on this implementation detail.  Its output is byte-for-byte
identical to that of the previous implementation, and installation was
tested on all platforms.

We continue to use sun.security.pkcs.PKCS10 on the servers, as it
provides a greater range of functionality than this commit implements,
does not require shipping a native library, and since we can control the
server JRE, we can afford this shortcut.

Change-Id: I96f2646b6c611bf965561f55fa0291a5a3273549