nginx: disallow weak ciphers and keyexchanges
https://www.ssllabs.com/ssltest/ tells me that nginx on SP allows
insecure key exchanges.  The solution is to tell nginx to only use
strong ciphersuites.  It turns out that this behavior would be the
default; however, for reasons I do not understand, the puppet config we
used to ship makes nginx weaker.

Preferring RC4 is a recommendation from:
http://hynek.me/articles/hardening-your-web-servers-ssl-ciphers/
based on the fact that block ciphers in CBC mode keep getting attacked
via side channels.

Change-Id: I68b4984c4eeaceeed0939f6ec88d1eac0346acfb