AbstractChunker: fix resource leak
We never close the ResettableFileInputStream, which means the JVM holds
a file handle to that chunk temp file.  On Linux and OSX, this is fine,
but on Windows, that means that when FileEncodingBuffer.close() is
called, the backing file can't be deleted, so it gets added to the
"delete on exit" pool and remains on disk, taking up space.

Instead, we should simply release the chunk file handle when we're done
using it, which shoudl let the deletion happen immediately.

Change-Id: Icdce97ab55114c753bf329fd75b3f04865285bbc