Send business activation email for existing users:
Previously, if a user exists, nothing will happen if the user attempts to signup AeroFS for Business. After this change, an email will be sent but with slightly different content than the activation email sent for new users (see RequestToSignUppEmailler.

I also changed the /business/signup URL to /business/activate, to avoid confusion with account signup.