Remove organization's size + move StripeCustomerID to server lib from base lib.
Why the org size field is not needed:

o the current size of an organization can be computed by counting the users belonging to that org.
o in the future when we have laddered pricing plans, plan ids instead of org sizes should be saved in the db. this is because we will test various pricing models, and new tests should not affect existing organizations.

Change-Id: Id12d4a37aae376b6fbed67949e16d70713ed9511