Avoid unnecessary I/O in RAAU
If the target branch has a content hash that matches that
of the incoming object we should avoid unnecessary I/O as
well as useless DB writes.

This also prevents the creation of redundant versions,
and is especially important for aliasing as previously
aliasing large hierarchies would result in duplicating
them fully on all devices.

Change-Id: I260ea451a2384cb30a751c2131a4831144df8056