Add new device management service.
Interface:
    GET http://<host>:9019/last_seen/<did>
    GET http://<host>:9019/polling_interval

Example:
    GET http://sp.aerofs.com:9019/last_seen/155aa68bc35e426f9535619fb1d89d14
    GET http://sp.aerofs.com:9019/polling_interval

Where the polling interval is specified in seconds.

The last_seen page will return a 500 if the DID is malformed, a 404 if the
device is not found in the redis db (i.e. if the device has never been seen
online) and the last seen time in epoch seconds otherwise.

Default 5 minute granularity.

Currently hosting this on SP and deploying via the old monolithic deploy
mechanism. I'm open to changes in this area.

This interface is also described in the src/devman/README.txt file.

Change-Id: I9eba86a23bd25ce2aa0991110971c1b1330190e0