Fix uudecode invocation on ArchLinux
Turns out newer versions of uudecode don't like receiving *any* other
arguments if you pass a -o flag to uudecode.  Madness, but it's keeping
our ArchLinux users from being able to install (and could well cause a
similar on other distros in due time).

This patch redirects the file to stdin instead of passing it as an
argument to uudecode.

Change-Id: I871501f8ed13c91f396ffb2ac121ce08757f726c