Add IP address to device management service.
To accomplish this I also added some stuff to the verkehr admin REST
API:
    topic=>connection
    connection=>ip_address

The first is not strictly needed here, but will be useful in the future.
(I had originally added it because I thought it was needed in this case but I
was wrong. Leaving it here anyway as it might be useful.)

Also, I did some misc cleanup in the device management service (renames, and
a new more RESTful API after discussion with Peter). See the README for
details.

Change-Id: Ie8f079964a62d205057fd2ae5aa66d89468c4462