Fixed Finder being restarted on each update.
aerofs-1816
The cause of the issue was: when a client update occurrs, the client compares its Finder extensions checksum against the one stored in the database to decide if it needs to re-install Finder extensions. However, the client did not store the original checksum in some cases causing the comparisons to fail. The client then re-installs the Finder extensions and kill all Finder windows.

Also removed IOSUtil.getShellExtensionChecksum().

Change-Id: I449094bfbe6787ed7edcca6821685c0ba4d1a9d5