Avoid unnecessary IO
Backward-compatible change to the Core proto

When a device requests a CONTENT component from a peer, it
now advertises its own (MASTER) content hash, if available
(and below a reasonnable size threshold).

When a device replies to such a call, it compares its own
own content hash (for whichever branch it is actually
about to send) to the one advertised in the RPC (if any)
If the two hashes match, the reply will not contain any
actual content but instead a single is_content_same bit

When receiving a reply with the is_content_same bit set,
the device making the RPC can simply apply the received
version which requires much less cpu cycles and I/O than
the regular path.

This optimization is particularly useful in aliasing
situations as both devices end up computing content
hashes as part of the aliasing process anyway and
the opportunity to reduce disk bandwith and network
bandwidth leads to significant speed improvements.

A possible future improvement along the same lines
would be to detect when the peer making the call is
likely to discard the reply and ask for a hash to be
computed. In such cases, automatically computing the
content hash before sneding the reply would save some
more bandwidth and two round trips.

Change-Id: I7a992f2a2df46a30b7562d80983778ba9f3c0bd0