Send vk notifications last when joining
CRL updates can be long enough that clients receive the vk notifications
and pull ACL from SP before the transaction is committed, which prevents
the ACL changes from propagating properly as the client will not pull
again until a restart or another notification.

NB: a more comprehensive fix would require the vk notifications to be sent
outside of the transactions (probably not practical) or sp being able to
reject/delay ACL pulls during an ACL-modifying transaction or the client
being smart enough to retry in such situations.

Change-Id: I2b61438ae1432c9ad2df5b5d6f2a6ced807a117d