Update middleman PKI and bundled OpenJDK to match
Apparently Java 7's PKIX doesn't like CA certificates without basicConstraints:
CA:TRUE and critical.  While the rest of the client code uses SunX509 (which is
a bit more lax with enforcement), the updater uses URL.openConnection(), which
will use the platform default provider.

On Java 6, this is SunX509, so things would work, but on Java 7, it's PKIX.

The previous MITM cert was also well-formed in this manner, so that's why this
wasn't a problem before.

Change-Id: Icce5a6272f20fb05704d16b660436007bdcb2153