if admin invites a new user to team, the user automatically accepts the team invitation on sign-up + misc Web UI improvements:
Each team invitation can optionally attach a signup code. When an admin invites a new user to the team,
the signup code is attached to the team invitation. When a new user signs up using the sign up code,
the system automatically accept the team invitation that is attached with the code.

Change-Id: If47d0e76e3758d416e8aeffb95ebed93ae7e5b79