Store defects in a more easily explorable/contained manner.
Some users produce a ridiculous amount of log data, and there's not an easy way
to delete just these logs.  This commit restructures the defects folder to also
be organized by userid and device id.

It also creates symbolic links for backwards compatibility with other tools
(e.g. getdefects) in the previous location.

Change-Id: I5f3b66431a935285555988e1c6dfa24a700cd459