Speed up seed file creation
The larger a db, the worst aliasing would be so the more
important it is to be able to create a seed file for it.

Using DirectoryService.walk_ is just too slow because it
fetches whole OAs and CAs for every object and pollutes
the cache during the walk.

Using a lower-level walk primitive (a more informative
variant of getChildren_) yields a significant speed
improvement (~7x for 316k objects)

Also attach seed file to a store to prepare multiroot
support.

Change-Id: I07c54243e32e3f3c22b998999da68bc81ed70df6