Bootstrap server application.
This commit is really just packaging related. We now can create a
bootstrap debian that contains a startup script that parses an input file
in /opt/bootstrap/bootstrap.tasks and executes the tasks as defined by
the tasks input file. The format of the tasks file is one task name per line,
where the task name is just a string, for example, "verkehr".

The vm baking infrastructure will add this tasks file at some point in
the near future to the open stack baked vms depending on what services
are bundled in the vm application.

A sample verkehr task is provided but the implementation is not finished
yet.

Change-Id: I27bc33f8b9eb8db8344a69cfef57920c436b3ff9