TimeToSync: beginning of client implementation
Notice that none of this code will execute (on purpose). Most of the wiring
is in place to start recording time-to-sync, but TimeToSyncCollector *shouldn't*
be instantiated anywhere, and thus none of the listeners will notify this plugin
yet. The memory management isn't yet implemented, nor is the proper
testing/dumping/uploading of the histogram complete.
This commit is intended to break down the work into smaller parts.

Change-Id: I9fe63e71974f3467d68270427498c7432b7cd869