Fixed a bug in the Windows shell extension that made desktop icons flicker
Root cause: we were using SHChangeNotify(SHCNE_ASSOCCHANGED...) to refresh
the shell which is unnecessarily strong and refreshes everything. Switching
to SHChangeNotify(SHCNE_UPDATEDIR... refreshes only the root anchor.

That being said, there's an even deeper root cause which is that the
shellext is getting disconnected from the gui very often and then
reconnects automatically. Upon reconnection, we clear the cache, which was
triggering the refresh. It is still unclear why the shellext is getting
disconnected so often, but it has to do with Windows unloading it from
memory when its not needed.

Change-Id: I4a226836a4346346ddedbdeac33909852c1ea6e5