Clean up dead blocks on change if history disabled
AMENDED: added _ suffix to method names.
AMENDED: Also removed a TODO comment that no longer applies.

If version-history is disabled on a block storage system, schedule
the dead-block cleaner after any transaction where we deref
blocks. Effectively this means on the apply and delete paths,
we schedule removeDeadBlocks() to run after the transaction commits.

Previously backend storage would only be cleaned when a store is deleted.

This checkin adds unit tests that check for backend deletion calls
to match the deref'ed blocks. Also a small private change to the unit
test, reverting a signature change to store(...).

Change-Id: I5cf36f88ac279784c49fa478ba96c743d764164a
Task: aerofs-2004