Multiroot syncdet tests
This test suite needs two actors and the first actor
MUST be a TeamServer configured to use linked storage.

NB: due to a limitation in the python sp API, it is currently
not easy to obtain the list of shared folders for a Team Server
user. This is required by some tests so in the meantime a simple
workaround is to use the same account for the TeamServer and the
regular client

Relevant actor settings for actor 0 in config.yaml:
 aero_app_root: /home/aerofstest/syncdet/deploy/approot.multiuser/
 details:
    team_server: true
    storage_type: LINKED

Change-Id: I1899bb7ca3cc84f88ad130b3da1f038b62cfe0d0