Check for Windows-reserved file names
In OSUtilWindows, when we check if a filename could be valid,
disallow certain reserved file names that will be forbidden:
CON, PRN, NUL, COM1-9, etc., as well as filenames made up
of only dots. Reworked existing code to use a couple of
simple regexes.

Adds a tiny unit test for some simple filename patterns, positive and
negative.

This resolves support issue Support-818 which is affecting at least
a couple of current users.

Change-Id: Idaf039fae970fe3298b74f08098c7b1ebe5d1400