Revert cache/gzip proxy ordering
Revert change in previous commit fixing S3MagicChunk check

The order used implies that chunks are still compressed in the
cache. This was not the original intent but it has been around
for a while and reverting it would require adding a DPUT to
clear the cache. This order trades cpu cycles for disk space
which may not be optimal but at this time it is highly unlikely
that decompressing cached chunks turn out to be a bottleneck
given the other profrormance issues we're observing.

Change-Id: I81f869dd78fb7b74662cf6b2625966cd03dbcd6d