Smooth SP upgrade path for external sharing
SP should not advertise ACL entries marked as external to
clients that cannot treat them with the required amount of
deference. Instead of enforcing that restriction through
a bump in the SP protocol version, these entries are by
default filtered out. The caller can indicate its ability
to deal with such entries by using a different RPC with
the exact same signature that does not filter entries.

This change does NOT break backward compatibility

Change-Id: Ice1b89b0ac3e0904618ba886aec27ef3b4a73223