Catch yet more Windows-illegal filenames.
This fixes the regex to correctly test for the Group Of Illegal Filenames
with file-extensions: LPT1.txt, CON.whatever, etc. Note that
"....txt" is _legal_, hence the fancy grouping in the regex.

The unit test is updated as well. Still hoping this will contribute to
a fix for issue SUPPORT-742, or at least fix our logging thereof.

AMENDED with changes suggested by Drew in over-the-shoulder review;
added a comment explaining the Fancy Regex Magic(tm), and added
positive test cases for variations on CLOCK$.

Change-Id: I71604062349c4d1a0b8762ef2b9abdb9876fb594