Kill daemon on failure to add watch
Use special ExitCode and catch it in DefaultDaemonMonitor to
inform the user of the problem.

NB: A bunch of Windows user seem to hit this problem when
the parent directory of the root anchor is not writable
due to a bug in JNotify that should be fixed in a separate
commit.

Change-Id: I88c6ebda9f0e95a50bd8547214083f8027a0d991