Don't scan unreadable directories in ScanSession
In ScanSession, we treat a directory that exists but is unreadable as an
exception case, and attempt to revisit it on the next pass. This causes
a ton of ExNotDir noise for affected users.

This change adds a check for canRead() before we push new subdirectory
entries onto the stack. canRead() is augmented with a check for a crazy
Windows situation; see the FIXME in the code for some justification.

Two known causes of unreadable directories, though there are certainly more:
 - an OSX/linux folder with no read permission for the current user
 - a Windows file symlink that points to a directory

This resolves issue AER-1965 (ScanSession throws exception ExNotDir for some resources)

Tests are updated accordingly.

AMENDED: Clarified wacky logic per HB's feedback. Thanks!
AMENDED: git error (PEBCAK).

Change-Id: I469b0f07819dd168bcc4b409be79b45105622810