Don't use deleteOnExit for temp files.
If we do, we'll leak them into the static filesToDelete Set that will grow
without bound over the lifetime of the daemon, eventually leading to OOM.

The only place we actually used this parameter is in AbstractChunker,
where we'd create a temp file for the file encoding buffer.  If we simply
clean this file up properly after use, we are much less likely to OOM the
Team Server.

Change-Id: I1fe4d17a2f9241faed94bc9236522056fb707495