AbstractChunker: avoid unnecessary calls to BSB
Every file chunk gets passed to the appropriate BlockStorageBackend, regardless
of whether that chunk is already referenced by the BlockStorageDatabase.  As a
result, throwing a 1GB file of all zeros onto S3 results in 256 HEAD requests,
where only one is needed.

This significantly reduces the time needed to chunk and upload files with
repeated chunks to S3.

Change-Id: I2772803ed198cab4518c1c4c224baab745012e4e