One-way migration on TeamServer
With the availability of Linked storage on TS it is necessary to
add support for locally initiated migration (i.e. in reaction to
a filesystem move across shared folder boundary).

Remotely-initiated migration is still ignored as the uniqueness
of admitted OID across locally present stores cannot be garanteed
in a multiuser configuration.

NB: it might be possible to support a subset of remotely
initiated migration by using copy semantics instead of move
but it is outside the scope of this commit

Change-Id: I1d640c3b0d0304eaa73ef290a6f8a07831a1dd91